--- application/tool/tool-base/base-exception/src/main/java/pers/posse/tool/ws/ExternalException.java
@@ -1,28 +1,0 @@
-package pers.posse.tool.ws;
-
-import pers.posse.tool.ExceptionContext;
-import pers.posse.tool.ToolException;
-
-/**
- * Created by posse on 17-7-31.
- */
-public class ExternalException extends ToolException {
-    public ExternalException() {
-    }
-
-    public ExternalException(String message) {
-        super(message);
-    }
-
-    public ExternalException(Throwable cause) {
-        super(cause);
-    }
-
-    public ExternalException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ExternalException(String message, Throwable cause, ExceptionContext context) {
-        super(message, cause, context);
-    }
-}
