--- application/tool/tool-base/base-exception/src/main/java/pers/posse/tool/ws/WsAuthenticationException.java
@@ -1,22 +1,0 @@
-package pers.posse.tool.ws;
-
-/**
- * Created by posse on 17-7-31.
- */
-public class WsAuthenticationException extends RuntimeException{
-    private static final long serialVersionUID = 1L;
-
-    private String responseCode;
-
-    public WsAuthenticationException() {
-
-    }
-
-    public WsAuthenticationException(String responseCode) {
-        this.responseCode = responseCode;
-    }
-
-    public String getResponseCode() {
-        return responseCode;
-    }
-}
