+++ application/tool/tool-core/public-api-service/src/main/java/per/posse/tool/ws/enums/OperationType.java
@@ -1,0 +1,8 @@
+package per.posse.tool.ws.enums;
+
+/**
+ * Created by posse on 17-7-25.
+ */
+public enum OperationType {
+    CREATE, RETRIEVE, UPDATE, DELETE;
+}
