--- application/tool/tool-core/public-api-service/src/main/java/pers/posse/tool/web/ws/enums/OperationDomain.java
@@ -1,8 +1,0 @@
-package pers.posse.tool.web.ws.enums;
-
-/**
- * Created by posse on 17-7-25.
- */
-public enum OperationDomain {
-    STUDENT, TEACHER, COURSE, USER;
-}
