--- application/tool/tool-core/public-api-service/src/main/java/pers/posse/tool/web/ws/enums/OperationType.java
@@ -1,8 +1,0 @@
-package pers.posse.tool.web.ws.enums;
-
-/**
- * Created by posse on 17-7-25.
- */
-public enum OperationType {
-    CREATE, RETRIEVE, UPDATE, DELETE;
-}
