--- application/tool/tool-core/public-api-service/src/main/java/pers/posse/tool/web/ws/xml/DomainAttribute.java
@@ -1,105 +1,0 @@
-package pers.posse.tool.web.ws.xml;
-
-import domain.enums.Gender;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
-
-/**
- * Created by posse on 17-7-25.
- */
-@XmlRootElement(name = "attribute")
-@XmlAccessorType(XmlAccessType.FIELD)
-public class DomainAttribute {
-
-    private Long id;
-
-    private String name;
-
-    private Integer age;
-
-    private Gender gender;
-
-    private String idNum;
-
-    private String address;
-
-    private String mobile;
-
-    private String apiName;
-
-    private String apiPassword;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    public Gender getGender() {
-        return gender;
-    }
-
-    public void setGender(Gender gender) {
-        this.gender = gender;
-    }
-
-    public String getIdNum() {
-        return idNum;
-    }
-
-    public void setIdNum(String idNum) {
-        this.idNum = idNum;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getMobile() {
-        return mobile;
-    }
-
-    public void setMobile(String mobile) {
-        this.mobile = mobile;
-    }
-
-    public String getApiName() {
-        return apiName;
-    }
-
-    public void setApiName(String apiName) {
-        this.apiName = apiName;
-    }
-
-    public String getApiPassword() {
-        return apiPassword;
-    }
-
-    public void setApiPassword(String apiPassword) {
-        this.apiPassword = apiPassword;
-    }
-}
