--- application/tool/tool-core/public-api-service/src/main/java/pers/posse/tool/web/ws/xml/ExternalOperateEntityResponse.java
@@ -1,32 +1,0 @@
-package pers.posse.tool.web.ws.xml;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
-
-/**
- * Created by posse on 17-7-19.
- */
-@XmlRootElement(name = "externalOperateEntityResponse")
-@XmlAccessorType(XmlAccessType.FIELD)
-public class ExternalOperateEntityResponse {
-    private String msg = "Insert Success.";
-
-    private DomainAttribute attribute;
-
-    public String getMsg() {
-        return msg;
-    }
-
-    public void setMsg(String msg) {
-        this.msg = msg;
-    }
-
-    public DomainAttribute getAttribute() {
-        return attribute;
-    }
-
-    public void setAttribute(DomainAttribute attribute) {
-        this.attribute = attribute;
-    }
-}
