+++ application/tool/tool-core/web/note.bak/codeExplanation/hibernate/InitHibernate.java
