+++ application/tool/tool-core/web/note.bak/config/application.properties.bak
@@ -1,0 +1,31 @@
+#################################
+### database dbcp properties
+#################################
+database.dbcp.testOnBorrow=false
+database.dbcp.testWhileIdle=true
+database.dbcp.timeBetweenEvictionRunsMillis=30000
+database.dbcp.minEvictableIdleTimeMillis=30000
+database.dbcp.numTestsPerEvictionRun=5
+database.dbcp.removeAbandoned=true
+database.dbcp.removeAbandonedTimeout=60
+database.dbcp.logAbandoned=true
+database.driver.className=com.mysql.jdbc.Driver
+#database.driver.className=org.postgresql.Driver
+
+database.school.read.url = jdbc:mysql://192.168.3.14:3306/school
+#database.school.read.url = jdbc:postgresql://192.168.3.14:5432/otmsopenapi
+database.school.read.username = posse
+database.school.read.password = posse
+database.school.read.initial.size = 1
+database.school.read.max.active = 50
+database.school.read.max.idle = 16
+database.school.read.maxWaitMillis = 5000
+
+database.school.write.url = jdbc:mysql://192.168.3.14:3306/school
+#database.school.write.url = jdbc:postgresql://192.168.3.14:5432/otmsopenapi
+database.school.write.username = posse
+database.school.write.password = posse
+database.school.write.initial.size = 1
+database.school.write.max.active = 50
+database.school.write.max.idle = 16
+database.school.write.maxWaitMillis = 5000
