--- application/tool/tool-core/web/src/main/java/pers/posse/tool/servlet/ServletDemo.java
@@ -1,8 +1,0 @@
-package pers.posse.tool.servlet;
-
-/**
- * Created by posse on 17-7-18.
- */
-public class ServletDemo {
-    
-}
