+++ application/tool/tool-core/web/src/main/resources/META-INF/config/cxf/webService.xml
@@ -1,0 +1,52 @@
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:jaxrs="http://cxf.apache.org/jaxrs"
+       xmlns:util="http://www.springframework.org/schema/util"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans
+	http://www.springframework.org/schema/beans/spring-beans-4.0.xsd
+	http://cxf.apache.org/jaxrs http://cxf.apache.org/schemas/jaxrs.xsd
+    http://www.springframework.org/schema/util
+    http://www.springframework.org/schema/util/spring-util-4.0.xsd ">
+
+    <bean id="externalOperateEntityServiceImpl" class="per.posse.tool.ws.impl.ExternalOperateEntityServiceImpl" />
+    <bean id="encodingLoggingInInterceptor" class="pers.posse.tool.interceptor.EncodingLoggingInInterceptor" />
+    <bean id="outLoggingInterceptor" class="org.apache.cxf.interceptor.LoggingInInterceptor" />
+
+    <bean id="jsonProvider" class="org.apache.cxf.jaxrs.provider.json.JSONProvider">
+        <property name="schemaHandler" ref="schemaHolder"/>
+    </bean>
+
+    <bean id ="jaxbProvider" class ="org.apache.cxf.jaxrs.provider.JAXBElementProvider">
+        <property name="schemaHandler" ref="schemaHolder" />
+    </bean>
+
+    <bean id="schemaHolder" class="org.apache.cxf.jaxrs.utils.schemas.SchemaHandler">
+        <property name="schemas" ref="theSchemas" />
+    </bean>
+
+    <util:list id="theSchemas">
+        <value>classpath*:/schema/webService.xsd</value>
+    </util:list>
+
+    <jaxrs:server id="externalInsertEntityService" address="/">
+        <jaxrs:serviceBeans>
+            <ref bean="externalOperateEntityServiceImpl" />
+        </jaxrs:serviceBeans>
+
+        <jaxrs:providers>
+            <ref bean="jaxbProvider" />
+            <ref bean='jsonProvider' />
+        </jaxrs:providers>
+
+        <jaxrs:inInterceptors>
+            <ref bean="encodingLoggingInInterceptor"/>
+        </jaxrs:inInterceptors>
+        <jaxrs:outInterceptors>
+            <ref bean="outLoggingInterceptor"/>
+        </jaxrs:outInterceptors>
+        <jaxrs:extensionMappings>
+            <entry key="json" value="application/json" />
+            <entry key="xml" value="application/xml" />
+            <entry key="en" value="en-gb" />
+        </jaxrs:extensionMappings>
+    </jaxrs:server>
+</beans>
