+++ application/tool/tool-core/web/src/main/resources/META-INF/config/persistence/persistence.xml
@@ -1,39 +1,13 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <persistence xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="2.0" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd">
-<persistence-unit name="persistenceUnitWrite" transaction-type="RESOURCE_LOCAL">
+<persistence-unit name="persistenceUnit" transaction-type="RESOURCE_LOCAL">
         <provider>org.hibernate.ejb.HibernatePersistence</provider>
-        <!--<non-jta-data-source>schoolDataSourceRead</non-jta-data-source>-->
+        <non-jta-data-source>schoolDataSource</non-jta-data-source>
         <properties>
-            <!--<property name="hibernate.dialect" value="org.hibernate.dialect.PostgreSQL9Dialect"/>-->
             <property name="hibernate.dialect" value="org.hibernate.dialect.MySQL5Dialect"/>
-
-            <!-- value="create" to build a new database on each run; value="update" to modify an existing database;
-                value="create-drop" means the same as "create" but also drops tables when Hibernate closes;
-                value="validate" makes no changes to the database
-            -->
             <property name="hibernate.hbm2ddl.auto" value="update"/>
             <property name="hibernate.show_sql" value="true" />
             <property name="hibernate.format_sql" value="true" />
-
-
-            <property name="hibernate.jdbc.use_streams_for_binary" value="false"/>
-            <property name="hibernate.ejb.naming_strategy" value="org.hibernate.cfg.ImprovedNamingStrategy"/>
-            <property name="hibernate.connection.charSet" value="UTF-8"/>
-            <property name="hibernate.order_updates" value="true" />
-
-            <!-- 
-			<property name="hibernate.cache.use_second_level_cache" value="true"/>
-			<property name="hibernate.cache.provider_class" value ="net.sf.ehcache.hibernate.EhCacheProvider"/>
-			http://ehcache.org/documentation/2.4/toolUser-guide/hibernate#Downloading-and-Installing-Ehcache
-             -->
-            
-            <!-- Uncomment the following two properties for JBoss only -->
-            <!-- property name="hibernate.validator.apply_to_ddl" value="false" /-->
-            <!-- property name="hibernate.validator.autoregister_listeners" value="false" /-->
         </properties>
-        <class>pers.posse.tool.service.impl.domain.ToolUser</class>
-        <class>pers.posse.tool.service.impl.domain.Student</class>
-        <class>pers.posse.tool.service.impl.domain.Teacher</class>
-        <class>pers.posse.tool.service.impl.domain.Course</class>
     </persistence-unit>
 </persistence>
