+++ application/tool/tool-core/web/src/main/resources/META-INF/config/spring/applicationContext.xml
@@ -1,19 +1,66 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:aop="http://www.springframework.org/schema/aop"
        xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:tx="http://www.springframework.org/schema/tx"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:tx="http://www.springframework.org/schema/tx"
        xsi:schemaLocation="http://www.springframework.org/schema/beans
 	         http://www.springframework.org/schema/beans/spring-beans-4.0.xsd
-             http://www.springframework.org/schema/aop
-             http://www.springframework.org/schema/aop/spring-aop-4.0.xsd
 	         http://www.springframework.org/schema/context
 	         http://www.springframework.org/schema/context/spring-context-4.0.xsd
              http://www.springframework.org/schema/tx
 	         http://www.springframework.org/schema/tx/spring-tx-4.0.xsd ">
 
-    <import resource="classpath*:META-INF/config/cxf/webServices.xml" />
-    <import resource="classpath*:META-INF/config/spring/applicationContext_dataSource.xml" />
+    <import resource="classpath*:META-INF/config/cxf/webService.xml" />
+    <context:property-placeholder location="classpath*:META-INF/*.properties"/>
+    <context:component-scan base-package="pers.posse.tool.service.impl" />
+
+    <bean id="schoolDataSource" class="org.apache.commons.dbcp.BasicDataSource" destroy-method="close">
+        <property name="driverClassName"  value="${database.driver.className}" />
+        <property name="url" value="${database.school.read.url}" />
+        <property name="username" value="${database.school.write.username}" />
+        <property name="password" value="${database.school.write.password}" />
+    </bean>
+
+    <bean id="persistenceUnitManager" class="org.springframework.orm.jpa.persistenceunit.DefaultPersistenceUnitManager">
+        <property name="persistenceXmlLocations">
+            <list>
+                <value>classpath*:META-INF/config/persistence/persistence.xml</value>
+            </list>
+        </property>
+        <property name="dataSources">
+            <map>
+                <entry key="dataSource" value-ref="schoolDataSource"/>
+            </map>
+        </property>
+    </bean>
+
+    <bean id="entityManagerFactory" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
+        <property name="persistenceUnitName" value="persistenceUnit"/>
+        <property name="persistenceUnitManager" ref="persistenceUnitManager"/>
+        <property name="dataSource" ref="schoolDataSource" />
+        <property name="jpaVendorAdapter">
+            <bean id="jpaVendorAdapter" class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter">
+                <property name="database" value="MYSQL"/>
+            </bean>
+        </property>
+        <property name="jpaDialect">
+            <bean class="org.springframework.orm.jpa.vendor.HibernateJpaDialect"/>
+        </property>
+        <property name="jpaPropertyMap">
+            <map>
+                <entry key="hibernate.ejb.interceptor.session_scoped">
+                    <value>pers.posse.tool.interceptor.HibernateSessionInterceptor</value>
+                </entry>
+            </map>
+        </property>
+    </bean>
+
+    <bean id="transactionManager" class="org.springframework.orm.jpa.JpaTransactionManager">
+        <property name="entityManagerFactory" ref="entityManagerFactory"/>
+    </bean>
+
+    <tx:annotation-driven transaction-manager="transactionManager" mode="aspectj" />
+
 </beans>
