+++ application/tool/tool-core/web/src/main/webapp/index.html
@@ -4,6 +4,6 @@
     </head>
 
     <body>
-        <a>测试</a>
+        <a>Posse 测试</a>
     </body>
 </html>
